Create Microsoft & Space LEAF Corp non-profit
This is for all around the globe, artist, creators, actors, actresses, men, women, children alike this is something new to be able to have your creative flow and know that you're not being spied on, and that your contributions aren't just making you money, but also allowing us all to grow and expand as a global planet not just a people and not just every nation individual. We all get to grow together at our own pace in time this is all optional, but this is a dream I had to bring something new to people and have fun with life again.